refactor(daemon): extract file-level helpers and evaluation hooks from daemon.ts
Pure-move extraction of ~60 file-level helpers, constants, types and error
classes into nine modules under src/daemon/ (turn-types, tool-classification,
github-hook-coords, loop-guard-scope, text, webchat, helpers, constants,
evaluation-types). Previously exported symbols stay re-exported from daemon.ts
so no importer or test changes.

The evaluation-harness method cluster moves into DaemonEvaluationHooks
(src/evaluation/daemon-hooks.ts) behind a narrow DaemonEvaluationHost deps
interface; Daemon keeps one-line delegates with unchanged public signatures.

daemon.ts shrinks 23,852 → 22,285 lines with no behavior change.
